+++ src/activateMockDebug.ts
@@ -1,0 +1,138 @@
+/*---------------------------------------------------------
+ * Copyright (C) Microsoft Corporation. All rights reserved.
+ *--------------------------------------------------------*/
+
+'use strict';
+
+import * as vscode from 'vscode';
+import { WorkspaceFolder, DebugConfiguration, ProviderResult, CancellationToken } from 'vscode';
+import { MockDebugSession } from './mockDebug';
+import { FileAccessor } from './mockRuntime';
+
+export function activateMockDebug(context: vscode.ExtensionContext, factory?: vscode.DebugAdapterDescriptorFactory) {
+
+	context.subscriptions.push(
+		vscode.commands.registerCommand('extension.mock-debug.runEditorContents', (resource: vscode.Uri) => {
+			vscode.debug.startDebugging(undefined, {
+				type: 'mock',
+				name: 'Run Editor Contents',
+				request: 'launch',
+				program: resource.fsPath
+			}, {
+				//noDebug: true
+			});
+		}),
+		vscode.commands.registerCommand('extension.mock-debug.debugEditorContents', (resource: vscode.Uri) => {
+			vscode.debug.startDebugging(undefined, {
+				type: 'mock',
+				name: 'Debug Editor Contents',
+				request: 'launch',
+				program: resource.fsPath
+			});
+		}),
+		vscode.commands.registerCommand('extension.mock-debug.showAsHex', (variable) => {
+			vscode.window.showInformationMessage(`${variable.container.name}: ${variable.variable.name}`);
+		})
+	);
+
+	context.subscriptions.push(vscode.commands.registerCommand('extension.mock-debug.getProgramName', config => {
+		return vscode.window.showInputBox({
+			placeHolder: "Please enter the name of a markdown file in the workspace folder",
+			value: "readme.md"
+		});
+	}));
+
+	// register a configuration provider for 'mock' debug type
+	const provider = new MockConfigurationProvider();
+	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider('mock', provider));
+
+	// register a dynamic configuration provider for 'mock' debug type
+	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider('mock', {
+		provideDebugConfigurations(folder: WorkspaceFolder | undefined): ProviderResult<DebugConfiguration[]> {
+			return [
+				{
+					name: "Dynamic Launch",
+					request: "launch",
+					type: "node",
+					program: "${file}"
+				},
+				{
+					name: "Another Dynamic Launch",
+					request: "launch",
+					type: "node",
+					program: "${file}"
+				},
+				{
+					name: "Mock Launch",
+					request: "launch",
+					type: "node",
+					program: "${file}"
+				}
+			];
+		}
+	}, vscode.DebugConfigurationProviderTriggerKind.Dynamic));
+
+	if (!factory) {
+		factory = new InlineDebugAdapterFactory();
+	}
+	context.subscriptions.push(vscode.debug.registerDebugAdapterDescriptorFactory('mock', factory));
+	if ('dispose' in factory) {
+		context.subscriptions.push(factory);
+	}
+
+	// override VS Code's default implementation of the debug hover
+	/*
+	vscode.languages.registerEvaluatableExpressionProvider('markdown', {
+		provideEvaluatableExpression(document: vscode.TextDocument, position: vscode.Position): vscode.ProviderResult<vscode.EvaluatableExpression> {
+			const wordRange = document.getWordRangeAtPosition(position);
+			return wordRange ? new vscode.EvaluatableExpression(wordRange) : undefined;
+		}
+	});
+	*/
+}
+
+class MockConfigurationProvider implements vscode.DebugConfigurationProvider {
+
+	/**
+	 * Massage a debug configuration just before a debug session is being launched,
+	 * e.g. add all missing attributes to the debug configuration.
+	 */
+	resolveDebugConfiguration(folder: WorkspaceFolder | undefined, config: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration> {
+
+		// if launch.json is missing or empty
+		if (!config.type && !config.request && !config.name) {
+			const editor = vscode.window.activeTextEditor;
+			if (editor && editor.document.languageId === 'markdown') {
+				config.type = 'mock';
+				config.name = 'Launch';
+				config.request = 'launch';
+				config.program = '${file}';
+				config.stopOnEntry = true;
+			}
+		}
+
+		if (!config.program) {
+			return vscode.window.showInformationMessage("Cannot find a program to debug").then(_ => {
+				return undefined;	// abort launch
+			});
+		}
+
+		return config;
+	}
+}
+
+export const workspaceFileAccessor: FileAccessor = {
+	async readFile(path: string) {
+		const uri = vscode.Uri.parse(path);
+		const bytes = await vscode.workspace.fs.readFile(uri);
+		const contents = Buffer.from(bytes).toString('utf8');
+		return contents;
+	}
+};
+
+class InlineDebugAdapterFactory implements vscode.DebugAdapterDescriptorFactory {
+
+	createDebugAdapterDescriptor(_session: vscode.DebugSession): ProviderResult<vscode.DebugAdapterDescriptor> {
+		return new vscode.DebugAdapterInlineImplementation(new MockDebugSession(workspaceFileAccessor));
+	}
+}