+++ src/web/extension.ts
@@ -1,0 +1,14 @@
+/*---------------------------------------------------------
+ * Copyright (C) Microsoft Corporation. All rights reserved.
+ *--------------------------------------------------------*/
+
+import * as vscode from 'vscode';
+import { activateMockDebug } from '../activateMockDebug';
+
+export function activate(context: vscode.ExtensionContext) {
+	activateMockDebug(context);
+}
+
+export function deactivate() {
+	// nothing to do
+}